refactor lookup tables into a package
