+++ tests/test_interpolation.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from vivarium.interpolation import (
+from vivarium.framework.lookup.interpolation import (
     Interpolation,
     Order0Interp,
     check_data_complete,
